djangooauth2tinyset/client/requirements.txt
Django<1.10
#django-allauth
django-storages
boto
psycopg2
pytz
pip-tools
pip
django-storages
coverage
django-nose
#django-oauth-toolkit
#django-cors-headers
python-social-auth
